+++ VTK/Utilities/vtknetcdf/dim.c
@@ -2,7 +2,7 @@
  *  Copyright 1996, University Corporation for Atmospheric Research
  *      See netcdf/COPYRIGHT file for copying and redistribution conditions.
  */
-/* $Id: dim.c,v 1.1 2005-07-15 21:56:38 andy Exp $ */
+/* $Id: dim.c,v 1.2 2005-07-19 12:31:30 andy Exp $ */
 
 #include "nc.h"
 #include <stdlib.h>
@@ -99,7 +99,7 @@ find_NC_Udim(const NC_dimarray *ncap, NC_dim **dimpp)
   {
     /*EMPTY*/
   }
-  if(dimid >= ncap->nelems)
+  if((size_t)dimid >= ncap->nelems)
     return(-1); /* not found */
   /* else, normal return */
   if(dimpp != NULL)
@@ -137,7 +137,7 @@ NC_finddim(const NC_dimarray *ncap, const char *name, NC_dim **dimpp)
   {
     /*EMPTY*/
   }
-  if(dimid >= ncap->nelems)
+  if((size_t)dimid >= ncap->nelems)
     return(-1); /* not found */
   /* else, normal return */
   if(dimpp != NULL)
